Add missing things, re-export more things.
I missed a couple of Dev's additions to "the core code" when I pulled it
out of Birli and hyperdrive; these should now be available.

Moved the "precession" module under "pos" because that seems like more
of a positional thing. No minor version bump needed because the crate
re-exports the module anyway.

Remove static assertions on MWA coordinates; after chatting with Dev
and Greg, we agreed that these coordinates should live in this crate,
not mwalib.

Loosened up dependency versions; one day, ndarray 0.16 will be released,
which will use approx 0.5, which uses num-complex 0.4. No changes to
this crate should be needed to use the latest versions of these packages
when that day comes.