Don't label assert lines as partially covered.
This is a little controversial; technically, an assert is not totally
covered until it has passed and failed, but practically, we don't care
about the fail case. Asserts are used in tests and tests don't pass if
the asserts don't pass. More info here:
https://github.com/mozilla/grcov/issues/476

This commit labels lines like assert!, assert_eq!, debug_assert_eq!,
etc. as fully covered, as well as derived traits (e.g. Debug).

This commit *does not* label lines with error handling (i.e. ?) as fully
covered; this is more nuanced than the assert case.