Specify minimum dep versions.
A bunch of dependency versions were allowed but cause Marlu to fail to
compile. This commit doesn't fix *everything*, because some dependencies
are still allowing some of their dependencies to be broken. But things
are better now.

Add a test against the specified minimum Rust version and for CUDA. Also
fix the minimum Rust version not being right.