Use erfa-sys 0.2.0
The new version of erfa-sys is mostly quality of life for that repo, but
there are benefits for Marlu too. One is that a newer version of
ERFA (2.0.0) is statically compiled if that feature is enabled. Another
is that the new erfa-sys fixes some issues with the constants; Marlu has
been altered to take advantage of these.