Optimise uvfits writing.
The existing uvfits IO benchmarks improve from 15.762s to 4.3323s and
33.512s to 17.715s with this commit. The main difference is using raw
cfitsio calls throughout the uvfits writing code; the old code used it
as little as possible, opting for the fitsio crate convenience, but this
required the entire uvfits file to be written out effectively
twice (oops). There should also be a lot less copying and heap
allocations now because a buffer is stored and reused inside
UvfitsWriter. Also remove as many unwraps as possible and cleanups
previously-leaked CStrings.

The price of this performance gain is losing the Send + Sync traits on
the VisWritable trait. However none of the other MWA libraries were
using this, and we're unlikely to anyway; our writing functions are not
designed to write in parallel, but even then, parallel IO tends to make
things worse than just working in serial.

Change some other public things, like the function argument `obs_name`
from `Option<String>` to `Option<&str>`.

Bump the minimum-required version of Rust so we can use the nice
usize::abs_diff function. Also bump the version of the nightly compiler
used in coverage to use this function.