Make a new LmnRime type.
The `LMN::prepare_for_rime` function previously returned a new LMN, but
this was a bit misleading. This function now returns `LmnRime`, which
is (hopefully) distinct and can be converted back to LMN.

Also add `dot` functions on `LMN` and `LmnRime` to conveniently do a dot
product with a `UVW`.

Add tests for all aspects of LMN code.