Make Jones ABI stable and add approx impls.
`Jones` is now `#[repr(transparent)]`, which is important for safety
when doing thing like transmuting.

`approx` is now a Marlu feature, that when enabled, exposes trait
implementations `AbsDiffEq` and `RelativeEq` on all coordinate types.

Also derive `PartialEq` for a handful of coordinate types that had it
missing, as well as `Eq` for `Jones`.